Bugfix #5224
Prob: selection sometimes gave wrong edges/faces

Selection code for editmesh was still using a short for calculating the
distance of a projected vertex/edge/face from the mouse position. In
zoomed in cases that'll give overflows and unpredictable results.

It was fixed only half before... now all shorts are removed for distance
calculus.

+++ source/blender/include/editmesh.h
@@ -89,8 +89,8 @@ extern void translateflag(short flag, float *vec);
 extern int convex(float *v1, float *v2, float *v3, float *v4);
 
 /* ******************* editmesh_mods.c */
-extern EditEdge *findnearestedge(short *dist);
-extern EditVert *findnearestvert(short *dist, short sel);
+extern EditEdge *findnearestedge(int *dist);
+extern EditVert *findnearestvert(int *dist, short sel);
 
 /* ******************* editmesh_tools.c */
